Dockerfile: do not specify alpine version in nginx image
This is hard to reconcile with renovate. It will either try to lock the
same alpine version (with `docker` versioning), or try to update to
non-alpine versions (with `loose` versioning).

Not specifying the alpine version should make the `docker` versioning
useful while at the same time allowing alpine updates via the digest.